Changes in Phase Analysis
The temporal analysis notebook is now almost complete. It now includes a table of contents with the specific links, the Phase Analysis can now change the colors of the different signals and we now added two more parameters to tune in the phase analysis: the critical frequencies for the Butterworth filter and the order parameter also for the filter. TODO: I still have to fix the backend issue between ginput and ipywidgets. Whenever I use ipywidgets backend I cannot use ginpu.