Extra changes for release
Changes in Spatial and Temporal analysis to improve redeability and easy access inside the notebooks. Moreover, Spatial analysis now includes the LOI to be able to use a line of interest to project points. TO DO: make this in form of a widget or hide more the code for cleaner view of the notebook